requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
google-auth>=2.22.0 